google_cloud_storage/requirements.txt
# coding=utf-8
# Copyright 2023 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
#

requests==2.28.1
jwt==1.3.1
google-auth==2.15.0
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-secret-manager==2.13.0
google-cloud-storage==2.7.0